+++ docs/source/changelog.rst
@@ -13,6 +13,7 @@ Enhancements
 - better link and size finding for HTTP (#610, %99)
 - link following in Local (#608)
 - ReferenceFileSystem dev (#606, #604, #602)
+- Enable listings cache for HTTP filesystem (#560)
 
 Fixes